+++ apps/api/src/controllers/adminOrderController.js
@@ -2,48 +2,64 @@ import Order from '../models/Order.js';
 import User from '../models/User.js';
 import mongoose from 'mongoose';
 
-// 모든 주문 목록 조회 (관리자용)
-export const getAllOrders = async (req, res) => {
+// 일반 유저 목록 조회
+export const getUsers = async (req, res) => {
   try {
-    const { page = 1, limit = 10 } = req.query;
+    const page = parseInt(req.query.page) || 1;
+    const limit = parseInt(req.query.limit) || 15;
+    const skip = (page - 1) * limit;
 
-    const orders = await Order.find({})
-      .sort({ createdAt: -1 })
-      .limit(limit * 1)
-      .skip((page - 1) * limit)
-      .populate('userId', 'name email')
-      .populate('items.productId', 'productName productImg price');
+    // 현재 날짜 기준으로 이번 달의 시작과 끝 계산
+    const now = new Date();
+    const startOfMonth = new Date(now.getFullYear(), now.getMonth(), 1);
+    const endOfMonth = new Date(now.getFullYear(), now.getMonth() + 1, 0, 23, 59, 59, 999);
 
-    const total = await Order.countDocuments({});
+    // 총 일반 유저 수
+    const totalUsers = await User.countDocuments();
 
-    res.json({
-      orders: orders.map(order => ({
-        _id: order._id,
-        orderNumber: order.orderNumber,
-        userId: {
-          _id: order.userId._id,
-          name: order.userId.name,
-          email: order.userId.email
-        },
-        items: order.items,
-        totalPrice: order.totalPrice,
-        paymentMethod: order.paymentMethod,
-        createdAt: order.createdAt
-      })),
-      pagination: {
-        currentPage: parseInt(page),
-        totalPages: Math.ceil(total / limit),
-        totalOrders: total
+    // 이번 달 가입한 유저 수
+    const newUsersThisMonth = await User.countDocuments({
+      createdAt: {
+        $gte: startOfMonth,
+        $lte: endOfMonth
       }
     });
 
+    // 유저 리스트 (이름, 이메일, 가입일)
+    const users = await User.find()
+      .select('name email createdAt')
+      .sort({ createdAt: -1 })
+      .skip(skip)
+      .limit(limit);
+
+    // 응답 데이터 구성
+    const response = {
+      summary: {
+        totalUsers,
+        newUsersThisMonth
+      },
+      users: users.map(user => ({
+        name: user.name,
+        email: user.email,
+        createdAt: user.createdAt
+      })),
+      pagination: {
+        currentPage: page,
+        totalPages: Math.ceil(totalUsers / limit),
+        totalUsers,
+        hasNextPage: page < Math.ceil(totalUsers / limit),
+        hasPrevPage: page > 1
+      }
+    };
+
+    res.json(response);
   } catch (error) {
-    console.error('모든 주문 목록 조회 오류:', error);
+    console.error('유저 목록 조회 오류:', error);
     res.status(500).json({ message: '서버 오류가 발생했습니다.' });
   }
 };
 
-// 특정 사용자 주문 목록 조회 (관리자용)
+// 일반 유저 주문 목록 조회
 export const getUserOrders = async (req, res) => {
   try {
     const { userId } = req.params;
