+++ apps/api/src/controllers/eventController.js
@@ -1,8 +1,14 @@
+/* ------------------------------------------------------------
+ * File      : /src/controllers/eventController.js
+ * Brief     : 이벤트 관련 컨트롤러
+ * Author    : 송용훈
+ * Date      : 2025-08-15
+ * Version   : 
+ * History
+ * ------------------------------------------------------------*/
+
 import Event from '../models/Event.js';
 import { compressMulterFile } from '../utils/imageUtils.js';
-import fs from 'fs';
-import path from 'path';
-import os from 'os';
 import mongoose from 'mongoose'; // Added for reorderEvents
 
 // 이벤트 등록
@@ -78,13 +84,6 @@ export const createEvent = async (req, res) => {
           { maxWidth: 1200, maxHeight: 400, quality: 85 }, 
           'event-image'
         );
-
-        console.log('이벤트 이미지 압축 완료:', {
-          원본크기: `${compressionResult.original.sizeKB}KB`,
-          압축크기: `${compressionResult.compressed.sizeKB}KB`,
-          압축률: `${compressionResult.compressionRatio}%`,
-          절약공간: `${Math.round(compressionResult.savedSpace / 1024 * 100) / 100}KB`
-        });
 
         processedEventImg = compressionResult.compressed.base64;
         
@@ -270,13 +269,6 @@ export const updateEvent = async (req, res) => {
           'event-image'
         );
 
-        console.log('이벤트 이미지 압축 완료:', {
-          원본크기: `${compressionResult.original.sizeKB}KB`,
-          압축크기: `${compressionResult.compressed.sizeKB}KB`,
-          압축률: `${compressionResult.compressionRatio}%`,
-          절약공간: `${Math.round(compressionResult.savedSpace / 1024 * 100) / 100}KB`
-        });
-
         updateData.eventImg = compressionResult.compressed.base64;
         
       } catch (compressionError) {
@@ -362,75 +354,3 @@ export const deleteEvent = async (req, res) => {
 };
 
 
-
-// 이벤트 순서 변경 (드래그 앤 드롭용)
-export const reorderEvents = async (req, res) => {
-  try {
-    const { eventIds } = req.body;
-
-    console.log('이벤트 순서 변경 요청:', { eventIds });
-
-    // 1. 유효성 검증
-    if (!Array.isArray(eventIds) || eventIds.length === 0) {
-      return res.status(400).json({ 
-        message: '유효하지 않은 이벤트 ID 배열입니다.',
-        receivedData: { eventIds }
-      });
-    }
-
-    // 2. 모든 이벤트 ID가 유효한지 확인
-    const validEvents = await Event.find({ _id: { $in: eventIds } });
-    if (validEvents.length !== eventIds.length) {
-      return res.status(400).json({ 
-        message: '존재하지 않는 이벤트가 포함되어 있습니다.',
-        receivedCount: eventIds.length,
-        validCount: validEvents.length
-      });
-    }
-
-    // 3. 트랜잭션 시작
-    const session = await mongoose.startSession();
-    session.startTransaction();
-
-    try {
-      // 4. 모든 이벤트의 eventOrder를 새로운 순서로 업데이트
-      for (let i = 0; i < eventIds.length; i++) {
-        await Event.findByIdAndUpdate(
-          eventIds[i],
-          { eventOrder: i + 1 },
-          { session }
-        );
-      }
-
-      // 5. 트랜잭션 커밋
-      await session.commitTransaction();
-
-      console.log('이벤트 순서 변경 성공:', {
-        updatedCount: eventIds.length,
-        newOrder: eventIds
-      });
-
-      res.json({
-        message: '이벤트 순서가 성공적으로 변경되었습니다.',
-        updatedCount: eventIds.length,
-        newOrder: eventIds
-      });
-
-    } catch (error) {
-      // 6. 에러 시 트랜잭션 롤백
-      await session.abortTransaction();
-      console.error('이벤트 순서 변경 트랜잭션 실패:', error);
-      throw error;
-    } finally {
-      // 7. 세션 종료
-      session.endSession();
-    }
-
-  } catch (error) {
-    console.error('이벤트 순서 변경 오류:', error);
-    res.status(500).json({
-      message: '서버 오류가 발생했습니다.',
-      error: error.message // 개발 중에만 사용
-    });
-  }
-};
