+++ apps/api/src/middlewares/upload.js
@@ -1,3 +1,12 @@
+/* ------------------------------------------------------------
+ * File      : /src/middlewares/upload.js
+ * Brief     : 파일 업로드 관련 라우트
+ * Author    : 송용훈
+ * Date      : 2025-08-10
+ * Version   : 
+ * History
+ * ------------------------------------------------------------*/
+
 import multer from 'multer';
 
 // 파일 필터링
