+++ apps/api/src/utils/logValidator.js
@@ -1,3 +1,12 @@
+/* ------------------------------------------------------------
+ * File      : /src/utils/logValidator.js
+ * Brief     : 로그 검증 함수
+ * Author    : 송용훈
+ * Date      : 2025-09-01
+ * Version   : 
+ * History
+ * ------------------------------------------------------------*/
+
 export function validateLogs(data) {
   const { logs } = data;  // 구조 분해 할당
